+++ pages/_app.js
@@ -1,8 +1,15 @@
-import '../styles/globals.css'
-import '../styles/index.css'
+import { Analytics } from "@vercel/analytics/react";
+
+import "../styles/globals.css";
+import "../styles/index.css";
 
 const MyApp = ({ Component, pageProps }) => {
-  return <Component {...pageProps} />
-}
+  return (
+    <>
+      <Component {...pageProps} />
+      <Analytics />
+    </>
+  );
+};
 
-export default MyApp
+export default MyApp;