+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [v1.1.0] - `2023-04-11`
+
+- Upgraded dependencies
+- Fixed warning affected on newer dependencies
+- Added Vercel Analytics
+- Added gitignore for yarn PnP
+
 ## [v1.0.4] - `2022-08-02`
 
 - Removed API function serverless for temporary
@@ -23,6 +30,7 @@
 
 - Launched on first time with 3 features: cv, games, and donate
 
+[v1.1.0]: https://github.com/buildingwatsize/resume/releases/tag/v1.1.0
 [v1.0.4]: https://github.com/buildingwatsize/resume/releases/tag/v1.0.4
 [v1.0.3]: https://github.com/buildingwatsize/resume/releases/tag/v1.0.3
 [v1.0.2]: https://github.com/buildingwatsize/resume/releases/tag/v1.0.2